# Conflicts:
#	.gitignore
#	.idea/misc.xml
#	Final2024.iml
#	out/production/Final2024/Rooms/room1.txt
#	out/production/Final2024/Rooms/room2.txt
#	src/Entities/Enemy.java
#	src/Entities/Entity.java
#	src/Entities/Player.java
#	src/Entities/Projectile.java
#	src/Entities/RangedEnemy.java
#	src/Entities/ShortMeleeEnemy.java
#	src/GamePanel.java
#	src/Items/IntelligencePickup.java
#	src/Items/Melee/MeleeWeapon.java
#	src/Items/Ranged/RangedWeapon.java
#	src/Items/WeaponData/BasicSpear.txt
#	src/Items/WeaponData/BasicSword.txt
#	src/Managers/CameraManager.java
#	src/Managers/EntityManager.java
#	src/Structure/Grid.java
#	src/Structure/Hitbox.java
#	src/Structure/Room.java
#	src/Structure/Vector2F.java
#	src/Universal/Camera.java

+++ src/Structure/Line.java
@@ -43,26 +43,26 @@ public class Line {
         x = h.getLeft();
         y = (slope * x + b);
         if (h.getTop() <=  y && y <= h.getBottom()) {
-            intersectionPoints.add(new Vector2F(x, y));
+            intersectionPoints.add(new Vector2F((int) x, (int) y));
         }
 
         x = h.getRight();
         y = (slope * x + b);
         if (h.getTop() <= y && y <= h.getBottom()) {
-            intersectionPoints.add(new Vector2F(x, y));
+            intersectionPoints.add(new Vector2F((int) x, (int) y));
         }
 
         y = h.getTop();
         x = ((y - b) / slope);
         if (h.getLeft() <= x && x <= h.getRight()) {
-            intersectionPoints.add(new Vector2F(x, y));
+            intersectionPoints.add(new Vector2F((int) x, (int) y));
 
         }
 
         y = h.getBottom();
         x = ((y - b) / slope);
         if (h.getLeft() <= x && x <= h.getRight()) {
-            intersectionPoints.add(new Vector2F(x, y));
+            intersectionPoints.add(new Vector2F((int) x, (int) y));
 
         }
